+++ packages/sdk/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.7.11] - 2026-03-23
+
+_Dependency bump only._
+
 ## [0.7.9] - 2026-03-17
 
 _Version bump only._